riscv: Implement Kendryte K210 SoC clock structure
Add a clock provider driver for the Kendryte K210 SoC. This new
driver "kendryte,k210-clk", implements the full clock structure of
the SoC. Since it is required for the correct operation of the SoC,
this driver is compiled automatically when the SOC_KENDRYTE option is
selected.

With this change, the k210-sysctl driver is turned into a simple
platform driver which triggers populating its child nodes.

The clock structure implemented and many of the coding ideas for the
driver come from the work on the K210 support by Sean Anderson for thei
u-boot project.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>