riscv: Implement Kendryte K210 SoC clock structure
Add a clock provider driver for the Kendryte K210 SoC. This new
driver, k210-sysclk, implements the full clock structure of the SoC.
Since it is required for the correct operation of the SoC, this driver
is compiled automatically when the SOC_KENDRYTE option is selected.
The k210-sysctl driver is turned into a simple platform provider which
triggers the initialization of its child nodes drivers.

Device tree bindings for this driver are documented in
Documentation/devicetree/bindings/clock/kendryte,k210-sysctl.txt.

The clock structure implemented and many of the coding ideas for the
driver come from u-boot clock driver for the K210.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>