riscv: Add Kendryte K210 SoC clock driver
Add a clock provider driver for the Kendryte K210 SoC. This new
driver with compatible string "kendryte,k210-clk", implements the full
clock structure of the K210 SoC. Since it is required for the correct
operation of the SoC, this driver is automatically selected for
compilation when the SOC_KENDRYTE option is selected.

With this change, the k210-sysctl driver is turned into a simple
platform driver which enables its power bus clock and triggers
populating its child nodes. This driver retains the SOC early
initialization code, but this code now relies on the new function
k210_clk_early_init() provided by the new clk-k210 driver. This
function declaration is done using the new soc/kendryte/k210-sysctl.h
header file.

The clock structure implemented and many of the coding ideas for the
driver come from the work by Sean Anderson on the Kendryte K210 support
for the U-Boot project.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>