riscv: Enable Kendryte K210 reboot
The kendryte K210 SoC soft reset register can be used to soft reset
the board. Supporting this feature is trivial to implement using the
syscon-reboot generic driver. Enable this by modifying the sysctl
driver to execute devm_of_platform_populate() so that all children
nodes of the sysctl device node defined in the device tree, which
include the syscon-reboot compatible reboot node, are probed.

To facilitate the kernel configuration for this feature, add a
configuration option dependent on SOC_KENDRYTE in
arch/riscv/Kconfig.socs.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>