of: Fix gpios supplier parsing
The DesignWare gpio-dwapb GPIO driver ("snps,dw-apb-gpio" or
"apm,xgene-gpio-v2" compatible string) defines the now deprecated
property "snps,nr-gpios" to specify the number of GPIOs available
on a port. However, if this property is used in a device tree, its
"-gpios" suffix causes the generic property supplier parsing code to
interpret it as a cell reference when properties are parsed in
of_link_to_suppliers(), leading to an error messages such as:

OF: /soc/bus@50200000/gpio-controller@50200000/gpio-port@0: could not
find phandle

As this deprecated property is still present in many device trees, fix
this problem by manually defining a parse_gpios() instead of using the
DEFINE_SUFFIX_PROP() macro. This new parsing function issues a warning
and then ignores the deprecated property, skipping the phandle parsing
and thus avoiding the device tree parsing error.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>