riscv: Add Kendryte K210 clock debug information
Add a simple platform driver to Kendryte K210 clock driver to print
the clock structure initialized on boot. This feature depends on the
SOC_KENDRYTE_K210_CLK_DEBUG configuration option and is enabled at
runtime by adding the boolean property "debug" to the sysclk node of
the K210 device tree.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>