bus: Add simple-pm-bus optional clock handling
The simple-pm-bus driver bindings allow the declaration of an optional
bus clock. If this clock is defined by a platform device tree, it should
be enabled so that all devices under the bus operate correctly.
Otherwise, the clock infrastructure may simply disable the clock as it
may not have any reference.

Fix this by grabbing and enabling the bus clock if it is defined in the
device tree.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>