spi: dw: Fix locking around rx buffer
Move the test on the completion of a message reception from
interrupt_transfer() into dw_reader() so that the completion
evaluation can be executed under the spin lock. With this, dws->rx*
fields are all protected under the spinlock, which is consistent with
the setup done by dw_spi_transfer_one(). This fix also simplifies the
code in interrupt_transfer().

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>