spi: dw: Improve dw_reader()
On slow systems, relying on interrupts for signaling the presence of
bytes to read in the master FIFO can be too slow and result on many FIFO
overrun errors. This can be observed on the Kendryte K210 SoC for large
IOs to an SD card. Improve dw_reader() to reduce the occurence of these
overrun errors by adding an extra loop around the DW_SPI_DR read loop.
The allows catching data that was added to the FIFO by the device while
the previous loop was executing, removing the interrupt overhead.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>