dt-bindings: fix sifive gpio properties
The sifive gpio IP block supports up to 32 GPIOs. Reflect that in the
interrupts property description and maxItems. Also add the standard
ngpios property to describe the number of GPIOs available on the
implementation.

Also add the "canaan,k210-gpiohs" compatible string for the Canaan K210
SoC which uses the Sifive gpio IP block. For this compatible string,
do not define the clocks property as required as the K210 SoC does not
have a software controllable clock for the Sifive gpio block.

Cc: Yash Shah <yash.shah@sifive.com>
Cc: Paul Walmsley <paul.walmsley@sifive.com>
Cc: Rob Herring <robh@kernel.org>
Cc: devicetree@vger.kernel.org
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>