scsi: pm8001: fix pm80xx_set_sas_protocol_timer_config()
All fields of the SASProtocolTimerConfig structure have the __le32 type.
As such, use cpu_to_le32() to initialize them. This change suppresses
many sparse warnings:

warning: incorrect type in assignment (different base types)
   expected restricted __le32 [addressable] [usertype] pageCode
   got int

Note that the check to cap the value of the STP_IDLE_TMO field is
removed as this files is initialized using the fixed (and small) value
defined by the STP_IDLE_TIME macro.

The pm8001_dbg() calls printing the values of the SASProtocolTimerConfig
structure fileds are changed to use le32_to_cpu() to present the values
in CPU endianness/human readable form.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>