scsi: libsas: fix sas_ata_qc_issue()
libata completion path take the port lock. So the lock should not be
dropped while in issuing context. Remove the spin_unlock()/spin_lock()
of the ata port from sas_ata_qc_issue(). Keeping the lock is fine as
submission is done with GFP_ATOMIC allocations that do not sleep.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>