PCI: endpoint: Add NVMe endpoint function driver
Add a Linux PCI Endpoint function driver to implement a PCIe NVMe
device. The nvme endpoint function driver implements a PCIe controller
which executes NVMe commands sent by the host using an NVMe target which
can be either a file or a block device. Some NVMe admin commands are an
exception to this basic implementation: the create submission queue and
create completion queue commands for the admin and IO queues are parsed
and executed by the driver directly to allow for mapping on the endpoint
memory the host defined queues in the host PCIe space.

When initialized, the nvme endpoint driver brings up an nvme controller
in the disabled state (CC.EN and CSTS.RDY are not set). When the host
enables the controller by setting CC.EN, the nvme endpoint driver
connects the nvme target with a local file or block device as backend
storage. This backend storage is exposed as the single namespace.

Workqueues are used to poll the controller registers and detect
controller enable/disable events, and to poll submission queues
doorbells to detect command submissions by the host.

Upon reception of a new command, the endpoint nvme driver parses the
command PRPs (or PRP list) to map and access host data buffers. The
execution of the commands on the nvme target side are done using a local
buffer which is transfered to or from the host based on the PRP mapping.

This driver code is based on an RFC submission by Alan Mikhak
<alan.mikhak@sifive.com> (https://lwn.net/Articles/804369/).

Co-developed-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>
Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>