PCI: endpoint: Add NVMe endpoint function driver
Add a Linux PCI Endpoint function driver to implement an nvme device
over the PCIe bus. The nvme endpoint function driver implements a
controller exposed over PCIe to be used by a host nvme driver. For the
execution of the nvme commands sent by the host, the nvme endpoint
function driver uses Linux nvme target code. The only exception to this
are the create submission queue and create completion queue commands for
IO queues. These commands are locally parsed as a PCI memory space needs
to be allocated and exposed to the host by the nvme function driver to
allow for exchanging nvme commands and and nvme completion entries.

When initialized, the nvme endpoint driver brings up an nvme controller
in the disabled state (CC.EN and CSTS.RDY are not set). When the host
enables the controller by setting CC.EN, the nvme endpoint driver
connects the nvme target with a local file or block device as backend
storage. This backend storage is exposed as the single namespace
provided by the controller.

A workqueue is used to poll the controller register (to detect
enable/disable events) and the submission queues (admin and IO queues).
Commands received are executed using the nvme target code and the
queue_response() callback executed by the target code upon a command
completion to allow the endpoint driver to insert in the appropriate
completion queue a completion entry for the completed command.

Blah... Describe some more later about mapping and trnasfers.

Signed-off-by: Alan Mikhak <alan.mikhak@sifive.com>
Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>