PCI: rockchip: Fix window mapping and address translation for endpoint
The RK3399 PCI EP core has 33 windows for PCIe space, here in the driver
up to 32 fixed size (1M) windows are used and pages are allocated and
mapped accordingly. The driver first uses a single window and allocate
space inside which causes translation issues (between CPU space and PCI
space) because a window can only have a single translation at a given
time, which if multiple regions are allocated inside will cause conflicts.
Fix this by making each window a single region of 1M, which will always
guarantee that the translation is not in conflict.

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>
Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>