pci: endpoint: Simplify epf test driver results print
In pci_epf_test_print_rate(), instead of open coding a reduction loop to
allow for a disivision by a 32-bits ns value, simply use div64_u64() to
calculate the rate. To match the printed unit of KB/s, this calculation
divides the rate by 1000 instead of 1024 (that would be KiB/s unit).

The format of the results printed by pci_epf_test_print_rate() is also
changed to be more compact without the double new line. dev_info() is
used instead of pr_info().

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>