PCI: rockchip: Do not advertize support for MSIX
The rockchip EP controller includes in the EP configuration space a PCI
capability entry for MSIX. However, the driver does not have support for
MSIX and it is impossible to declare "0" interrupts supported using the
MSIX capability structure. As a result, a host may be left beleiving
that the EP function supports MSIX but this support is missing.

Temporarily clear the capability ID of the MSIX capability entry to
avoid confusing the host.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>