PCI: rockchip: Do not advertize support for MSIX
The rockchip EP controller includes in the EP configuration space a PCI
capability entry for MSIX. However, the rockchip controller does not
support MSIX. This entry should thus not be present in the endpoint
capability list to avoid that a host tries to use this non-existent
capability.

Modify the capability list to skip the MSIX entry by pointing to the
entry following the MSIX entry from the previosu capability in the list
(the MSI capability entry).

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>