PCI: epf-test: Simply pci_epf_test_raise_irq()
Change the interface of the function pci_epf_test_raise_irq() to
directly pass a pointer to the struct pci_epf_test_reg defining the test
being executed. This avoids the need for grabbing this pointer with a
cast of the register bar and simplifies the call sites as the irq type
and irq numbers do not have to be passed as arguments.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>