PCI: rockchip: Fix rockchip_pcie_prog_ep_ob_atu()
Only using the size of a PCI mapping to determine the number of lower
bits of the host PCI address that must be remapped to the local CPU
address works only if the host PCI address is aligned to the mapping
size. For cases where the PCI address is randomly aligned, handling the
entire PCI address range from pci_addr to pci_addr + size -1 may require
additional bits of address to be considered.

Fix this by calculating num_pass_bits using an XOR of the PCI mapping
start and end (included) address as that gives us the highest
significant bit of the address range changing from the start address.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>