PCI: rockchip: Improve rockchip_pcie_ep_unmap_addr()
There is no need to loop over all regions to find the memory window used
to map an address. We can use rockchip_ob_region() to determine the
region index, together with additional checks.

Furthermore, the ob_region_map bitmap should also be checked to ensure
that we are unmapping a mapped address.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>