PCI: endpoint: Introduce pci_epc_map_size()
Some endpoint controllers impose requirements on the alignment of the
controller memory address that must be used to map an RC PCI address.
For instance, the rockchip endpoint controller uses the lower bits of
the controller memory address to define the RC PCI address lower bits
used for a mapping, resulting in the need to match these lower bits when
using controller memory obtained with pci_epc_mem_alloc_addr().

Introduce the function pci_epc_map_size() and the map_size endpoint
controller operation to allow endpoint function drivers to obtain the
size and the offset into a controller address region that must be used
to map an RC PCI address range.

The map_size controller operation is optional: controllers that do not
have any controller address alignement constraint for mapping a RC PCI
address do not need to implement the map_size operation. For such
controller, pci_epc_map_size() always return the specified size and an
offset of 0.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>