PCI: endpoint: Introduce pci_epc_map_size()
Some endpoint controllers have requirements on the alignment of the
controller physical memory address that must be used to map an RC PCI
address region. For instance, the rockchip endpoint controller uses
at most the lower 22 bits of a physical memory address region to define
the lower bits of an RC PCI address. For mapping a PCI address region of
size bytes starting from pci_addr, the exact number of address bits
used is the number of address bits changing in the address range
[pci_addr..pci_addr + size - 1].

For this example, this creates the following constraints:
1) The offset into the controller physical memory allocated for a
   mapping depends on the mapping size *and* the starting PCI address
   for the mapping.
2) A mapping size cannot exceed the controller windows size (1MB) minus
   the offset needed into the allocated physical memory.

Generecially handling these constraints in a PCI EP function driver is
not possible with the current interface as it only provides the ->align
field in struct pci_epc_features. Furthermore, this alignment is static
and does not depend on a mapping pci address and size.

Solve this by introducing the function pci_epc_map_size() and the
endpoint controller operation ->map_size to allow endpoint function
drivers to obtain the size and the offset into a controller address
region that must be used to map an RC PCI address region. The size
information provided by pci_epc_map_size() can be used as the size
argument for the function pci_epc_mem_alloc_addr(). The offset value
provided by pci_epc_map_size() can be used to determine the phys_addr
argument passed to the function pci_epc_map_addr().

The controller operation ->map_size is optional: controllers that do not
have any address alignement constraint for mapping a RC PCI address
region do not need to implement the this operation. For such
controllers, pci_epc_map_size() always return the mapping size as equal
to the requested size and an offset equal to 0.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>