PCI: rockchip: Implement the map_offset endpoint operation
The rockchip PCIe endpoint controller handles PCIe transfers addresses
by masking the programmed PCI address lower bits and using the same
amount of lower bits from the CPU address space used for the mapping.
This means that when preparing a PCI address mapping, an endpoint
function driver must use a correct offset into the CPU region allocated
with pci_epc_mem_alloc_addr() in order for the transfers to reach the
correct PCI address on the RC.

Implement the map_offset endpoint controller operation to allow this.
The offset is calculated by determining the number of lower bits
changing over the range of PCI address of a mapping. That is, the offset
computation depends on both the start PCI address and the size of the
mapping.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>