PCI: rockchip: Implement the map_size endpoint controller operation
The rockchip PCIe endpoint controller handles PCIe transfers addresses
by masking the lower bits of the programmed PCI address and using the
same number of lower bits from the CPU address space used for the
mapping.

This means that when preparing a PCI address mapping, an endpoint
function driver must use a correct offset into the CPU region allocated
with pci_epc_mem_alloc_addr() in order for the transfers to reach the
correct PCI address on the RC, and a sufficient amount of memory must be
allocated to accomodate for this offset.

Implement the map_size endpoint controller operation to allow this
transparently from the endpoint function drivers. The offset is
calculated by determining the number of lower bits changing over the
range of RC PCI addresses of a mapping. That is, the offset computation
depends on both the start RC PCI address and the size of the mapping.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>