PCI: endpoint: Use pci_epf_mem_map/unmap() in test function driver
Modify the endpoint test driver to use the functions pci_epf_mem_map()
and pci_epf_mem_unmap() for the read, write and copy tests. For each
test case, the transfer (dma or mmio) are executed in a loop to ensure
that partial mappings returned by pci_epf_mem_map() are correctly
handled.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>