PCI: rockchip: Handle PERST signal in endpoint mode
Currently, the Rockchip PCIe endpoint controller driver does not handle
PERST signal, which prevents detecting when link training should
actually be started and also does not allow for defining link up/down
EPC events. This however can be supported using the controller ep_gpio
(active high).

Modify the endpoint rockchip driver to get the ep_gpio and its
associated interrupt which is serviced using a threaded IRQ with the
function rockchip_pcie_ep_perst_irq_thread() as handler.

This handler function notifies a link down event corresponding to the RC
side asserting the PERST signal using pci_epc_linkdown() when the gpio
is high. Once the gpio value goes down, corresponding to the RC
de-asserting the PERST signal, link training is started. The polarity
of the gpio interrupt trigger is changed from high to low after the RC
asserted PERST, and conversely changed from low to high after the RC
de-asserts PERST.

Given that the rk3399 does not generate interrupts when the link status
changes, link training and link up notification is implemented using a
delayed work to poll for the completion of link training. This polling
is done as explained in the rk3399 technical reference manual:
  Insure link training completion and success by observing link_st field
  in PCIe Client BASIC_STATUS1 register change to 2'b11. If both side
  support PCIe Gen2 speed, re-train can be Initiated by asserting the
  Retrain Link field in Link Control and Status Register. The software
  should insure the BASIC_STATUS0[negotiated_speed] change to "1", that
  indicates re-train to Gen2 successfully.

The link training delayed work is scheduled when the RC de-asserts PERST
and stopped if the RC asserts PERST. When the link training work
completes and detects a stable link, a link up notification is issued to
the function driver using pci_epc_linkup().

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>