PCI: rockchip: Improve endpoint link training process
Currently, the Rockchip PCIe endpoint controller driver exhibits a
race condition between the link initialization and the EPF driver
starting operation (e.g. polling registers). This race manifests itself
with random SError NMIs crashing the kernel when the EPF accesses PCI
addresses too early. to avoid this, We need to first ensure that the
link is fully initialized, that is, that the link training is completed
before allowing the function driver to proceed.

This patch adds support for setting up an IRQ which triggers when the
host (RP side) asserts the link PERST signal, signaling that the
root-complex power and reference clocks are stable. Using this IRQ,
the link training status is probed for completion to detect a stable
link. The link training completion is then signaled to the function
driver using the linkup notifier.

Similarly, a link down can be detected when the hosts deasserts the
PERST signal, which results in an interrupt and notification of the
endpoint function driver.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>